requirements.txt
torch==2.2.0
torchvision==0.17.0
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.4
streamlit==1.32.0
langchain==0.1.12
langchain-community==0.0.29
langchain-openai==0.1.0
openai==1.14.0
faiss-cpu==1.8.0
sentence-transformers==2.5.1
scikit-learn==1.4.1
matplotlib==3.8.3
plotly==5.19.0
python-dotenv==1.0.1
reportlab==4.1.0
pydantic==2.6.3
tqdm==4.66.2
requests==2.31.0
